+++ lib/src/models.dart
@@ -6,9 +6,22 @@ part of 'app_runner.dart';
 typedef InitializeBinding = WidgetsBinding Function();
 
 /// {@template InitializeFunctions}
-/// A callback to initialize your code. Called during splash display (if any)
+/// A callback to initialize your code.
+/// Responsible for initializing custom methods, such as `Firebase`, `Sentry`, etc.
 /// {@endtemplate}
-typedef InitializeFunctions = FutureOr<void> Function(WidgetsBinding binding);
+typedef InitializeFunctions<T> = FutureOr<T> Function(WidgetsBinding binding);
+
+/// {@template RunnerBuilder}
+///  A builder that builds a widget based on a snapshot.
+/// Child can be null, is a cached version of the widget that does not depend on the snapshot.
+///
+/// This builder must only return a widget and should not have any side effects as it may be called multiple times.
+/// {@endtemplate}
+typedef RunnerBuilder<T> = Widget Function(
+  BuildContext context,
+  AsyncSnapshot<T?> snapshot,
+  Widget? child,
+);
 
 /// {@template OnError}
 /// The [onError] function is used both to handle asynchronous errors by overriding [ZoneSpecification.handleUncaughtError] in [zoneSpecification], if any, and to handle errors thrown synchronously by the call to [body].
@@ -19,62 +32,77 @@ typedef OnError = void Function(Object error, StackTrace stackTrace);
 /// Custom Builder for error handling in debug and profile mode
 /// {@endtemplate}
 typedef ErrorWidgetBuilder = Widget Function(
-    BuildContext context, FlutterErrorDetails errorDetails);
+  BuildContext context,
+  FlutterErrorDetails errorDetails,
+);
 
 /// {@template RunnerConfiguration}
-/// Runner Configuration, takes in itself:
-/// - [widgetConfig] is responsible for configuring launched widgets and handling errors in widgets;
-/// - [zoneConfig] responsible for the zone configuration in which your application will be launched;
-/// - [initializeBinding] is responsible for initializing your [WidgetsBinding]
-/// - [preInitializeFunctions] responsible for initializing custom methods, such as `Firebase` or `Sentry`.
+/// Runner Configuration
 /// {@endtemplate}
 @immutable
-class RunnerConfiguration {
-  /// {@macro RunnerConfiguration}
-  const RunnerConfiguration({
-    required this.widgetConfig,
-    required this.zoneConfig,
-    this.preInitializeFunctions,
-  });
+abstract class RunnerConfiguration {
+  /// Runner Configuration
+  /// - [widgetConfig] is responsible for configuring launched widgets and handling errors in widgets;
+  const factory RunnerConfiguration({
+    required WidgetConfiguration widgetConfig,
+  }) = _RunnerConfiguration;
+
+  /// Runner Configuration Guarded
+  /// - [widgetConfig] is responsible for configuring launched widgets and handling errors in widgets;
+  /// - [zoneConfig] responsible for the zone configuration in which your application will be launched;
+  const factory RunnerConfiguration.guarded({
+    required WidgetConfiguration widgetConfig,
+    required ZoneConfiguration zoneConfig,
+  }) = _RunnerConfigurationGuarded;
+
+  const RunnerConfiguration._(this.widgetConfig);
 
   /// {@macro WidgetConfiguration}
   final WidgetConfiguration widgetConfig;
+}
+
+class _RunnerConfiguration extends RunnerConfiguration {
+  /// {@macro RunnerConfiguration}
+  const _RunnerConfiguration({
+    required WidgetConfiguration widgetConfig,
+  }) : super._(widgetConfig);
+}
+
+class _RunnerConfigurationGuarded extends RunnerConfiguration {
+  /// {@macro RunnerConfiguration}
+  const _RunnerConfigurationGuarded({
+    required WidgetConfiguration widgetConfig,
+    required this.zoneConfig,
+  }) : super._(widgetConfig);
 
   /// {@macro ZoneConfiguration}
   final ZoneConfiguration zoneConfig;
-
-  /// {@macro InitializeFunctions}
-  final InitializeFunctions? preInitializeFunctions;
 }
 
 /// {@template WidgetConfiguration}
 /// Widget configuration, takes in itself:
-/// - [app] widget of your application;
+/// - [child] widget of your application;
 /// - [onFlutterError] error handler for flutter widgets;
-/// - [splash] loading widget of your application, completely independent of the [app];
 /// - [errorBuilder] handler for handling error screen in debug and profile mode. Attention! This widget fragile, do not do it loaded his best done through [LeafRenderObjectWidget];
-/// - [releaseErrorBuilder] handler for handling error screen in release mode. Attention! This widget fragile, do not do it loaded his best done through [LeafRenderObjectWidget].
+/// - [releaseErrorBuilder] handler for handling error screen in release mode. Attention! This widget fragile, do not do it loaded his best done through [LeafRenderObjectWidget];
+/// - [initializeBinding] is responsible for initializing your [WidgetsBinding];
 /// {@endtemplate}
 @immutable
 class WidgetConfiguration {
   /// {@macro WidgetConfiguration}
   const WidgetConfiguration({
-    required this.app,
+    required this.child,
     required this.onFlutterError,
-    this.splash,
     this.errorBuilder,
     this.releaseErrorBuilder,
     this.initializeBinding,
   });
 
   /// Your application widget
-  final Widget app;
+  final Widget child;
 
   /// Called whenever the Flutter framework catches an error.
   final void Function(FlutterErrorDetails errorDetails) onFlutterError;
-
-  /// Your splash widget
-  final Widget? splash;
 
   /// {@macro ErrorWidgetBuilder}
   final ErrorWidgetBuilder? errorBuilder;
@@ -88,9 +116,9 @@ class WidgetConfiguration {
 
 /// {@template ZoneConfiguration}
 /// Zone configuration, takes in itself:
-/// - [onZoneError] zone error handler
-/// - [zoneValues] local data for a specific zone
-/// - [zoneSpecification] zone specification
+/// - [onZoneError] zone error handler;
+/// - [zoneValues] local data for a specific zone;
+/// - [zoneSpecification] zone specification;
 /// {@endtemplate}
 @immutable
 class ZoneConfiguration {